+++ src/Test/java/ru/javawebinar/topjava/UserTestData.java
@@ -1,5 +1,4 @@
 package ru.javawebinar.topjava;
-
 import ru.javawebinar.topjava.model.Role;
 import ru.javawebinar.topjava.model.User;