+++ src/Test/java/ru/javawebinar/topjava/repository/inmemory/InMemoryUserRepository.java
@@ -9,7 +9,7 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static ru.javawebinar.topjava.UserTestData.*;
+import static UserTestData.*;
 
 
 @Repository