+++ src/Test/java/ru/javawebinar/topjava/service/UserServiceTest.java
@@ -17,7 +17,7 @@ import ru.javawebinar.topjava.util.exception.NotFoundException;
 import java.util.List;
 
 import static org.junit.Assert.assertThrows;
-import static ru.javawebinar.topjava.UserTestData.*;
+import static UserTestData.*;
 
 @ContextConfiguration({
         "classpath:spring/spring-app.xml",